feat: Rewards Breakdown Mobile
